Replaced the problematic tag spinner in the toolbar with a simple TextView that mimics a spinner
